+++ aliases.config.js
@@ -20,7 +20,7 @@ const aliases = {
 	"@pages": "src/pages",
 	"@utils": "src/utils",
 	"@plugins": "src/plugins",
-	"@server": "src/server",
+	"@serverMiddleware": "src/serverMiddleware",
 	"@store": "src/store",
 	"@static": `src/themes/${themeName}/static`,
 	"@theme": `src/themes/${themeName}`,
