+++ nuxt.config.js
@@ -56,7 +56,7 @@ module.exports = {
 		port: process.env.PORT || 4000,
 		host: process.env.HOST || "localhost",
 	},
-	serverMiddleware: ["@server/proxy"],
+	serverMiddleware: ["@serverMiddleware/proxy"],
 
 	router: {
 		middleware: [
