+++ package.json
@@ -9,7 +9,6 @@
 		"dev:storybook": "start-storybook -c stories/.storybook -p 4001",
 		"dev:docs": "vuepress dev docs",
 		"dev": "concurrently \"npm run dev:nuxt\" \"npm run dev:storybook\" \"npm run dev:docs\" -n \"nuxt,story,docs\" -c \"green,magenta,cyan\" -i false --kill-others-on-fail",
-		"build:legacy": "cd src/legacy-client && npx gulp",
 		"build:nuxt": "cross-env NODE_ENV=production nuxt build",
 		"build:docs": "vuepress build docs -d dist/docs",
 		"build:storybook": "build-storybook -c stories/.storybook -o dist/storybook",
