+++ src/middleware/links-fallback.js
@@ -1,4 +1,7 @@
-const routes = require("../server/routes");
+const vueRoutes = require("../serverMiddleware/routes.js");
+
+const isNuxtRoute = (url) =>
+	vueRoutes.some((regexString) => !!new RegExp(regexString).exec(url));
 
 export default async function(ctx) {
 	const { route } = ctx;
@@ -7,39 +10,14 @@ export default async function(ctx) {
 		return;
 	}
 
-	const segments = route.path.split("/");
-	const controllerName = segments[1];
-	const legacyRouteConfig = routes.find((r) => {
-		return typeof r === "object"
-			? r.route.includes(controllerName)
-			: r.includes(controllerName);
-	});
+	const useNuxt = isNuxtRoute(route.path);
 
-	// no matching route => use (default) error handler
-	if (!route.matched.length) {
-		return;
-	}
-	// no legacy route, but vue route found => use vue version
-	if (!legacyRouteConfig) {
-		return;
-	}
 	// prevent loop when ID not castable
 	if (window.location.pathname == route.path) {
 		return;
 	}
 
-	// route defined in vue and legacy client
-	// use legacy if not excluded
-	const useLegacy =
-		// no routes excluded
-		!(legacyRouteConfig || {}).routesExcluded ||
-		// or current url doesn't match any excluded url
-		(legacyRouteConfig.routesExcluded || []).every((excludedRoute) => {
-			const fullPath = "/" + legacyRouteConfig.route + excludedRoute;
-			return !fullPath.match(route.matched[0].regex);
-		});
-
-	if (useLegacy) {
+	if (!useNuxt) {
 		window.location = route.path;
 		// prevent rendering of nuxt during window load (switch to fallback)
 		return new Promise((resolve) => {
