--- src/server/routes.js
@@ -1,7 +1,0 @@
-// Whitelisted Vue routes
-// everything else will be forwarded to the legacy client
-
-module.exports = [
-	//
-	"^/news",
-];
