+++ startup.sh
@@ -2,7 +2,6 @@
 
 if [ ! -d "/app/dist" ]
 then
-	npm run build:legacy
 	npm run build:nuxt
 fi
